requirements.txt
Flask==3.0.2
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
PyPDF2
python-docx
pywin32 ; sys_platform == 'win32'
groq
python-dotenv
reportlab
pandas

